requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
pytest==8.1.1
PyYAML==6.0.1
selenium==4.24.0
webdriver-manager==4.0.2